Actualización

+++ src/main/java/com/esthetic/reservations/api/dto/InventoryDTO.java
@@ -10,9 +10,9 @@ import com.esthetic.reservations.api.model.Branch;
 public class InventoryDTO extends GenericModelDTO {
 	private String inventory_name;
 	
-	private Integer price;
+	private Double price;
 	
-	private Integer store;
+	private Long store;
 	
 	private String imagen;
 	
@@ -22,7 +22,7 @@ public class InventoryDTO extends GenericModelDTO {
 		super();
 	}
 
-	public InventoryDTO(Long id, String inventory_name, Integer price, Integer store, String imagen, Branch id_branch) {
+	public InventoryDTO(Long id, String inventory_name, Double price, Long store, String imagen, Branch id_branch) {
 		super(id);
 		this.inventory_name = inventory_name;
 		this.price = price;
@@ -39,19 +39,19 @@ public class InventoryDTO extends GenericModelDTO {
 		this.inventory_name = inventory_name;
 	}
 
-	public Integer getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(Integer price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 
-	public Integer getStore() {
+	public Long getStore() {
 		return store;
 	}
 
-	public void setStore(Integer store) {
+	public void setStore(Long store) {
 		this.store = store;
 	}
